Make Node.js the only runtime
Both the API and the federation worker had a Deno entry point and a
Node twin, and mise carried a task pair for each so either could be
started.  Compose still ran the Deno side.  Nothing was left to learn
from keeping both, so drop the Deno halves and rename the survivors:
main.node.ts becomes main.ts and so on for worker, instrument, logging,
worker-scheduler, and non-closing-writable.  The `:node` task suffixes
go with them.

Unpin Deno and stop installing its dependencies, drop the `check:deno`
and `test:deno` gates and their CI steps, and collapse codegen, check,
and test into one task each.  The production image no longer carries the
Deno binary or its module cache; a build confirms neither is present and
that the health check tasks still run inside it.

@sentry/deno loses its only importer with the Deno instrument entry, so
remove the dependency.  Tests that skipped under Deno, and the fixtures
written around the two-runtime split, now describe a single runtime.

The transport-error branches in logFilter.ts that match Deno-shaped
fetch failures stay: they are covered by tests, and dropping noise
suppression at the cutover would be a poor trade.  Their comments now
say where those shapes came from.

The Deno configuration files themselves are removed separately.

https://github.com/hackers-pub/hackerspub/issues/351
