+++ graphql/composition.test.ts
@@ -4,48 +4,31 @@ import test from "node:test";
 
 const readTextFile = (path: string | URL) => readFile(path, "utf8");
 
-for (const compositionRoot of ["main.ts", "worker.ts"]) {
-  test(`${compositionRoot} initializes LogTape after Sentry instrumentation`, async () => {
-    const source = await readTextFile(
-      new URL(compositionRoot, import.meta.url),
-    );
-    const instrumentImport = 'import "./instrument.ts";';
-    const loggingImport = 'import "./logging.ts";';
-
-    assertStringIncludes(source, instrumentImport);
-    assertStringIncludes(source, loggingImport);
-    assert(
-      source.indexOf(instrumentImport) < source.indexOf(loggingImport),
-      "Sentry instrumentation must run before LogTape configuration",
-    );
-  });
-}
-
 for (const [compositionRoot, role] of [
-  ["main.node.ts", "API"],
-  ["worker.node.ts", "worker"],
+  ["main.ts", "API"],
+  ["worker.ts", "worker"],
 ] as const) {
-  test(`the Node ${role} preloads Sentry before evaluating its module graph`, async () => {
+  test(`the ${role} preloads Sentry before evaluating its module graph`, async () => {
     const [source, tasks] = await Promise.all([
       readTextFile(new URL(compositionRoot, import.meta.url)),
       readTextFile(new URL("../mise.toml", import.meta.url)),
     ]);
-    const loggingImport = 'import "./logging.node.ts";';
-    const preload = `--import ./instrument.node.ts ${compositionRoot}`;
+    const loggingImport = 'import "./logging.ts";';
+    const preload = `--import ./instrument.ts ${compositionRoot}`;
 
     assertStringIncludes(source, loggingImport);
     assertStringIncludes(tasks, preload);
     assert(
-      tasks.indexOf("--import ./instrument.node.ts") <
+      tasks.indexOf("--import ./instrument.ts") <
         tasks.indexOf(compositionRoot),
-      `Node Sentry instrumentation must be preloaded before the ${role} entrypoint`,
+      `Sentry instrumentation must be preloaded before the ${role} entrypoint`,
     );
   });
 }
 
-test("the Node API owns unhandled rejection classification", async () => {
+test("the API owns unhandled rejection classification", async () => {
   const instrument = await readTextFile(
-    new URL("instrument.node.ts", import.meta.url),
+    new URL("instrument.ts", import.meta.url),
   );
 
   assertStringIncludes(
@@ -57,10 +40,10 @@ test("the Node API owns unhandled rejection classification", async () => {
 });
 
 for (const [compositionRoot, role] of [
-  ["main.node.ts", "API"],
-  ["worker.node.ts", "worker"],
+  ["main.ts", "API"],
+  ["worker.ts", "worker"],
 ] as const) {
-  test(`the Node ${role} bounds Sentry shutdown flushing`, async () => {
+  test(`the ${role} bounds Sentry shutdown flushing`, async () => {
     const source = await readTextFile(
       new URL(compositionRoot, import.meta.url),
     );
@@ -87,20 +70,14 @@ test("empty Sentry DSNs do not enable shared integrations", async () => {
 });
 
 test("the queue worker migrates legacy deliveries before listening", async () => {
-  for (const [compositionRoot, queueStart] of [
-    ["worker.ts", "federation.startQueue("],
-    ["worker.node.ts", "runWorkerRuntime("],
-  ] as const) {
-    const source = await readTextFile(
-      new URL(compositionRoot, import.meta.url),
-    );
-    const migration = "await migrateLegacyOutboxEvents(db);";
+  const source = await readTextFile(new URL("worker.ts", import.meta.url));
+  const migration = "await migrateLegacyOutboxEvents(db);";
+  const queueStart = "runWorkerRuntime(";
 
-    assertStringIncludes(source, migration);
-    assertStringIncludes(source, queueStart);
-    assert(
-      source.indexOf(migration) < source.indexOf(queueStart),
-      `legacy outgoing messages must migrate before ${compositionRoot} starts the queue`,
-    );
-  }
+  assertStringIncludes(source, migration);
+  assertStringIncludes(source, queueStart);
+  assert(
+    source.indexOf(migration) < source.indexOf(queueStart),
+    "legacy outgoing messages must migrate before worker.ts starts the queue",
+  );
 });
