+++ graphql/main.ts
@@ -1,100 +1,141 @@
-// Must be the first import — see instrument.ts for the rationale.
-import "./instrument.ts";
+// Sentry is initialized by the `--import ./instrument.ts` preload before
+// this module graph is evaluated. LogTape must then be configured before any
+// application resources are created.
 import "./logging.ts";
 
+import { getLogger, dispose as disposeLogging } from "@logtape/logtape";
 import {
-  getDenoEnvironment,
+  getProcessEnvironment,
   loadGraphqlApiConfig,
 } from "@hackerspub/runtime/config";
+import { isMain } from "@hackerspub/runtime/main";
 import {
   createRuntimeResources,
   FILE_SYSTEM_STORAGE_BASE_URL,
+  type RuntimeResources,
 } from "@hackerspub/runtime/resources";
+import * as Sentry from "@sentry/node-sdk";
+import { readFile } from "node:fs/promises";
+import process from "node:process";
 import { createGraphqlApiHandler } from "./api.ts";
+import {
+  combineRuntimeAndCloseErrors,
+  closeSequentially,
+} from "./lifecycle.ts";
 import { createYogaServer } from "./mod.ts";
-import assetlinks from "./static/.well-known/assetlinks.json" with { type: "json" };
+import {
+  createNodeHttpServer,
+  type NodeHttpServer,
+  waitForNodeHttpShutdown,
+} from "./node-http.ts";
 import metadata from "./package.json" with { type: "json" };
-const appleAppSiteAssociationJson = Deno.readTextFileSync(
-  new URL("./static/.well-known/apple-app-site-association", import.meta.url),
-);
 
-const yogaServer = createYogaServer();
-const allowFileKv = Deno.args.includes("--allow-file-kv");
-const resources = await createRuntimeResources(
-  loadGraphqlApiConfig(getDenoEnvironment(), { allowFileKv }),
-  metadata.version,
-  {
-    fileSystemBaseUrl: FILE_SYSTEM_STORAGE_BASE_URL,
-    federation: {
-      manuallyStartQueue: true,
-      // TODO: Revert to Fedify's default RFC 9421-first behavior once
-      // https://github.com/bonfire-networks/activity_pub/issues/8 is fixed
-      // and released. Keep this aligned with the queue worker.
-      firstKnock: "draft-cavage-http-signatures-12",
-    },
-  },
-);
-const handler = createGraphqlApiHandler({
-  resources,
-  yogaServer,
-  assetlinksJson: JSON.stringify(assetlinks),
-  appleAppSiteAssociationJson,
-});
+const logger = getLogger(["hackerspub", "graphql"]);
+const SENTRY_CLOSE_TIMEOUT = 2_000;
 
-// The federation inbox/outbox queue worker and the periodic news-score sweep
-// run in a separate process (`worker.ts`), not here: keeping that background
-// work off this event loop and DB pool is what stops it from starving
-// user-facing GraphQL requests into Caddy 504s (WEB-NEXT-1W).
-const startServer = () =>
-  Deno.serve({ port: 8080 }, (request, connectionInfo) =>
-    handler(request, connectionInfo),
-  );
-
-let runtimeFailed = false;
-let runtimeError: unknown;
-try {
-  const server = startServer();
+function registerShutdownSignals(controller: AbortController): () => void {
   const signals = ["SIGINT", "SIGTERM"] as const;
-  const registeredSignals = new Set<(typeof signals)[number]>();
-  const removeSignalListeners = () => {
-    for (const signal of registeredSignals) {
-      Deno.removeSignalListener(signal, shutdown);
+  const listeners = new Map<NodeJS.Signals, () => void>();
+  for (const signal of signals) {
+    const listener = () => {
+      if (controller.signal.aborted) return;
+      logger.info("Received {signal}; shutting down the GraphQL API.", {
+        signal,
+      });
+      controller.abort();
+    };
+    listeners.set(signal, listener);
+    process.once(signal, listener);
+  }
+  return () => {
+    for (const [signal, listener] of listeners) {
+      process.off(signal, listener);
     }
-    registeredSignals.clear();
+    listeners.clear();
   };
-  const shutdown = () => {
-    removeSignalListeners();
-    void server.shutdown();
-  };
+}
+
+export async function main(): Promise<void> {
+  const shutdownController = new AbortController();
+  const removeSignalListeners = registerShutdownSignals(shutdownController);
+  let resources: RuntimeResources | undefined;
+  let httpServer: NodeHttpServer | undefined;
+  let yogaServer: ReturnType<typeof createYogaServer> | undefined;
+  let runtimeError: unknown;
+
   try {
-    for (const signal of signals) {
-      Deno.addSignalListener(signal, shutdown);
-      registeredSignals.add(signal);
-    }
-    await server.finished;
+    const [assetlinksJson, appleAppSiteAssociationJson] = await Promise.all([
+      readFile(
+        new URL("./static/.well-known/assetlinks.json", import.meta.url),
+        "utf8",
+      ),
+      readFile(
+        new URL(
+          "./static/.well-known/apple-app-site-association",
+          import.meta.url,
+        ),
+        "utf8",
+      ),
+    ]);
+    const allowFileKv = process.argv.includes("--allow-file-kv");
+    resources = await createRuntimeResources(
+      loadGraphqlApiConfig(getProcessEnvironment(), { allowFileKv }),
+      metadata.version,
+      {
+        fileSystemBaseUrl: FILE_SYSTEM_STORAGE_BASE_URL,
+        federation: {
+          manuallyStartQueue: true,
+          // Keep this aligned with the queue worker.
+          firstKnock: "draft-cavage-http-signatures-12",
+        },
+      },
+    );
+    yogaServer = createYogaServer();
+    const handler = createGraphqlApiHandler({
+      resources,
+      yogaServer,
+      assetlinksJson,
+      appleAppSiteAssociationJson,
+    });
+    httpServer = createNodeHttpServer(handler);
+    const address = await httpServer.listen();
+    logger.info("The Node.js GraphQL API is listening on {hostname}:{port}.", {
+      hostname: address.address,
+      port: address.port,
+    });
+    await waitForNodeHttpShutdown(httpServer.server, shutdownController.signal);
+  } catch (error) {
+    runtimeError = error;
   } finally {
     removeSignalListeners();
   }
-} catch (error) {
-  runtimeFailed = true;
-  runtimeError = error;
+
+  let closeError: unknown;
+  try {
+    await closeSequentially([
+      () => httpServer?.close(),
+      () => yogaServer?.dispose(),
+      () => resources?.close(),
+      () => disposeLogging(),
+    ]);
+  } catch (error) {
+    closeError = error;
+  }
+
+  const finalError = combineRuntimeAndCloseErrors(runtimeError, closeError);
+  if (finalError != null) Sentry.captureException(finalError);
+  try {
+    await Sentry.close(SENTRY_CLOSE_TIMEOUT);
+  } catch (error) {
+    if (finalError != null) {
+      throw new AggregateError(
+        [finalError, error],
+        "The GraphQL API failed and Sentry could not be closed.",
+      );
+    }
+    throw error;
+  }
+  if (finalError != null) throw finalError;
 }
 
-let closeFailed = false;
-let closeError: unknown;
-try {
-  await resources.close();
-} catch (error) {
-  closeFailed = true;
-  closeError = error;
-}
-if (runtimeFailed) {
-  if (closeFailed) {
-    throw new AggregateError(
-      [runtimeError, closeError],
-      "The GraphQL server failed and its resources could not be closed.",
-    );
-  }
-  throw runtimeError;
-}
-if (closeFailed) throw closeError;
+if (isMain(import.meta)) await main();
