+++ scripts/smoke-standalone.ts
@@ -164,8 +164,8 @@ try {
       "--import",
       "temporal-polyfill/global",
       "--import",
-      "./instrument.node.ts",
-      "main.node.ts",
+      "./instrument.ts",
+      "main.ts",
     ],
     graphqlDirectory,
     { KV_URL: standaloneKvUrl },
@@ -218,8 +218,8 @@ try {
       "--import",
       "temporal-polyfill/global",
       "--import",
-      "./instrument.node.ts",
-      "worker.node.ts",
+      "./instrument.ts",
+      "worker.ts",
     ],
     graphqlDirectory,
     {
